# src/apps/apis/communities-api/swagger.yaml
openapi: 3.0.3
info:
  title: Pigeon-swarm Communities API
  version: 1.0.0
  description: Endpoints for private/public communities, members, text channels, and channel messages.
...
paths:
  /communities/:
    get:
      tags:
        - communities
      summary: List communities for the authenticated member
      parameters:
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      responses:
        '200':
          description: Communities where the authenticated identity is a member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunitiesResource'
    post:
      tags:
        - communities
      summary: Create a community
      description: The authenticated identity becomes the owner and first member. visibility defaults to private and cannot be changed later. Private communities require encrypted channel message payloads. Public communities require plaintext channel message payloads so messages can be indexed and searched.
      parameters:
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommunityBody'
      responses:
        '200':
          description: Community created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
...
  /communities/{communityId}:
    get:
      tags:
        - communities
      summary: Get a private community
      description: Only community members can read the community.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      responses:
        '200':
          description: Community found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
        '409':
          description: Authenticated identity is not a community member.
    patch:
      tags:
        - communities
      summary: Update private community profile
      description: Only the community owner can update the name, description, or banner CID.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchCommunityBody'
      responses:
        '200':
          description: Community updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
        '409':
          description: Authenticated identity is not the community owner.
...
  /communities/{communityId}/members:
    get:
      tags:
        - communities
      summary: List private community members
      description: Only community members can list members.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      responses:
        '200':
          description: Community members.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityMembersResource'
    post:
      tags:
        - communities
      summary: Invite an identity to a private community
      description: Only the community owner or members with create_invites can create direct membership invitations. The invited identity is not added to memberIds until they accept the invitation.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommunityMemberBody'
      responses:
        '200':
          description: Pending invitation created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityMembershipRequestResource'
        '409':
          description: Authenticated identity does not have create_invites.
  /communities/{communityId}/members/{identityId}/roles:
    put:
      tags:
        - communities
      summary: Replace roles assigned to a community member
      description: Only the owner or members with manage_roles can assign roles. The built-in everyone role is implicit and does not need to be assigned.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - in: path
          name: identityId
          required: true
          schema:
            type: string
          description: URL-encoded identity id.
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutCommunityMemberRolesBody'
      responses:
        '200':
          description: Member roles replaced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
        '409':
          description: Authenticated identity does not have manage_roles.
  /communities/{communityId}/bans:
    post:
      tags:
        - communities
      summary: Ban an identity from a community
      description: Requires ban_members. Banned identities cannot request to join or accept invite links.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommunityBanBody'
      responses:
        '200':
          description: Identity banned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
...
  /communities/{communityId}/roles:
    get:
      tags:
        - communities
      summary: List community roles and member assignments
      description: Only community members can list roles.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      responses:
        '200':
          description: Community roles and member role assignments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityRolesResource'
    post:
      tags:
        - communities
      summary: Create a community role
      description: Only the owner or members with manage_roles can create roles.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityRoleBody'
      responses:
        '200':
          description: Role created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityRoleResource'
        '409':
          description: Authenticated identity does not have manage_roles.
...
  /communities/{communityId}/invites:
    post:
      tags:
        - communities
      summary: Create a private community invite link token
      description: Only the community owner or a member with create_invites can create invite link tokens. The token is bearer-style; anyone with a valid token can accept it with their signed identity. The optional encryptedCommunityKey is an opaque client-encrypted blob. The backend stores and returns it but never receives the fragment secret or a clear community key.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommunityInviteBody'
      responses:
        '200':
          description: Invite token created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityInviteResource'
        '409':
          description: Authenticated identity is not the community owner or the invite body is invalid.
...
  /communities/{communityId}/channels/voice:
    post:
      tags:
        - communities
      summary: Create a private community voice channel
      description: Requires manage_channels. Voice channel presence is derived from active calls scoped to that channel.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommunityVoiceChannelBody'
      responses:
        '200':
          description: Voice channel created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityVoiceChannelResource'
        '409':
          description: Authenticated identity does not have manage_channels.
...
components:
  parameters:
    CommunityId:
      in: path
      name: communityId
      required: true
      schema:
        type: string
    XIdentityId:
      in: header
      name: X-Identity-Id
      required: true
      schema:
        type: string
    XTimestamp:
      in: header
      name: X-Timestamp
      required: true
      schema:
        type: string
    XSignature:
      in: header
      name: X-Signature
      required: true
      schema:
        type: string
  schemas:
    CommunityTextChannelResource:
      type: object
      required:
        - id
        - name
        - type
        - createdAt
        - permissions
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - text
        createdAt:
          type: integer
        permissions:
          $ref: '#/components/schemas/CommunityChannelPermissionsResource'
        threads:
          type: array
          description: Recent active threads for this text channel, newest activity first.
          items:
            $ref: '#/components/schemas/CommunityChannelThreadSummaryResource'
...
    CommunityVoiceChannelResource:
      type: object
      required:
        - id
        - name
        - type
        - createdAt
        - connectedIdentityIds
        - permissions
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - voice
        createdAt:
          type: integer
        permissions:
          $ref: '#/components/schemas/CommunityChannelPermissionsResource'
        connectedIdentityIds:
          type: array
          description: Unique identities joined to active calls for this voice channel with at least one connected participant lease for the same call and identity. Historical joined participants without a connected lease are excluded.
          items:
            type: string
    CommunityResource:
      type: object
      required:
        - id
        - networkId
        - ownerIdentityId
        - name
        - description
        - memberIds
        - bannedMemberIds
        - memberRoles
        - roles
        - textChannels
        - voiceChannels
        - visibility
        - discoverable
        - autoJoinEnabled
        - createdAt
      properties:
        id:
          type: string
        networkId:
          type: string
        ownerIdentityId:
          type: string
        name:
          type: string
        description:
          type: string
        discoverable:
          type: boolean
          description: Whether this private community appears in discovery search results.
        autoJoinEnabled:
          type: boolean
          description: Whether non-banned identities can join without owner approval.
        avatar:
          type: string
          description: Public IPFS CID for the community avatar/icon.
        banner:
          type: string
          description: Public IPFS CID for the community banner.
        memberIds:
          type: array
          items:
            type: string
        bannedMemberIds:
          type: array
          items:
            type: string
        memberRoles:
          type: array
          items:
            $ref: '#/components/schemas/CommunityMemberRolesResource'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/CommunityRoleResource'
        textChannels:
          type: array
          items:
            $ref: '#/components/schemas/CommunityTextChannelResource'
        voiceChannels:
          type: array
          items:
            $ref: '#/components/schemas/CommunityVoiceChannelResource'
        visibility:
          type: string
          enum:
            - private
            - public
        createdAt:
          type: integer
    CommunitiesResource:
      type: object
      required:
        - communities
      properties:
        communities:
          type: array
          items:
            $ref: '#/components/schemas/CommunityResource'
...
    CommunityMembershipRequestResource:
      type: object
      required:
        - id
        - communityId
        - creatorIdentityId
        - identityId
        - type
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        communityId:
          type: string
        creatorIdentityId:
          type: string
        identityId:
          type: string
        type:
          type: string
          enum:
            - invitation
            - request
        status:
          type: string
          enum:
            - accepted
            - declined
            - pending
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
    CommunityMembersResource:
      type: object
      required:
        - memberIds
      properties:
        memberIds:
          type: array
          items:
            type: string
    CommunityRolesResource:
      type: object
      required:
        - roles
        - memberRoles
      properties:
        roles:
          type: array
          items:
            $ref: '#/components/schemas/CommunityRoleResource'
        memberRoles:
          type: array
          items:
            $ref: '#/components/schemas/CommunityMemberRolesResource'
    CommunityRoleResource:
      type: object
      required:
        - id
        - name
        - permissions
        - builtIn
      properties:
        id:
          type: string
        name:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/CommunityPermission'
        builtIn:
          type: boolean
    CommunityMemberRolesResource:
      type: object
      required:
        - identityId
        - roleIds
      properties:
        identityId:
          type: string
        roleIds:
          type: array
          items:
            type: string
    CommunityChannelPermissionsResource:
      type: object
      required:
        - visibleRoleIds
      properties:
        visibleRoleIds:
          type: array
          description: Role ids that can see and access the channel. Use everyone for all members.
          items:
            type: string
    CommunityInviteResource:
      type: object
      required:
        - inviteToken
        - communityId
        - maxUses
        - uses
      properties:
        inviteToken:
          type: string
        communityId:
          type: string
        encryptedCommunityKey:
          $ref: '#/components/schemas/EncryptedCommunityInviteKey'
        expiresAt:
          type: integer
        maxUses:
          type: integer
        uses:
          type: integer
...
    PostCommunityBody:
      type: object
      required:
        - networkId
        - name
        - description
      properties:
        networkId:
          type: string
        name:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 500
        avatar:
          type: string
          description: Optional public IPFS CID for the community avatar/icon.
        banner:
          type: string
          description: Optional public IPFS CID for the community banner.
        discoverable:
          type: boolean
          default: true
          description: Set false to keep this community out of discovery search results.
        autoJoinEnabled:
          type: boolean
          default: false
          description: Set true to let any non-banned identity join without owner approval.
        visibility:
          type: string
          default: private
          description: Select private for encrypted channel messages or public for plaintext searchable channel messages. This value is immutable after creation.
          enum:
            - private
            - public
    PatchCommunityBody:
      type: object
      required:
        - name
        - description
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        name:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 500
        avatar:
          type: string
          description: Optional public IPFS CID. Omit it to remove the avatar.
        banner:
          type: string
          description: Optional public IPFS CID. Omit it to remove the banner.
        discoverable:
          type: boolean
          description: Set false to keep this community out of discovery search results.
        autoJoinEnabled:
          type: boolean
          description: Set true to let any non-banned identity join without owner approval.
    PostCommunityMemberBody:
      type: object
      description: The mutation signs a community_membership_request document of type invitation (creator = actor, identityId = invitee, status pending, updatedAt = createdAt). Its id is the first 24 hex chars of sha256(JSON.stringify([communityId, type, creatorIdentityId, identityId, createdAt])).
      required:
        - identityId
        - createdAt
        - mutation
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        createdAt:
          type: integer
          minimum: 0
        mutation:
          $ref: '#/components/schemas/SignedPublicMutation'
        identityId:
          type: string
    PostCommunityInviteBody:
      type: object
      description: The mutation signs a community_invite document in store requests whose id and token are base64url(sha256(JSON.stringify([communityId, creatorIdentityId, nonce]))). Signer is the creator.
      required:
        - nonce
        - createdAt
        - mutation
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        nonce:
          type: string
          minLength: 16
          maxLength: 128
        createdAt:
          type: integer
          minimum: 0
        mutation:
          $ref: '#/components/schemas/SignedPublicMutation'
        encryptedCommunityKey:
          $ref: '#/components/schemas/EncryptedCommunityInviteKey'
        expiresAt:
          type: integer
          description: Optional millisecond timestamp after which the invite cannot be accepted.
        maxUses:
          type: integer
          minimum: 1
          default: 1
...
    CommunityModerationLogBody:
      type: object
      description: Client-signed moderation log entry. The mutation signs {action, actorIdentityId, communityId, createdAt, details, id, scopeType 'community_moderation_log', target {id, type}} in store moderationLogs, recordId = id, sequence 0, signer = the acting moderator. id is the first 24 hex chars of sha256(JSON.stringify([communityId, actorIdentityId, action, target.type, target.id, createdAt])). Channel and role creation use createdAt as the input of the derived channel/role id.
      required:
        - createdAt
        - mutation
      properties:
        createdAt:
          type: integer
          minimum: 0
        mutation:
          $ref: '#/components/schemas/SignedPublicMutation'
    PostCommunityBanBody:
      type: object
      required:
        - identityId
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        identityId:
          type: string
        reason:
          type: string
...
    PostCommunityVoiceChannelBody:
      type: object
      required:
        - name
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        name:
          type: string
          maxLength: 80
...
    CommunityRoleBody:
      type: object
      required:
        - name
        - permissions
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        name:
          type: string
          maxLength: 50
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/CommunityPermission'
    PutCommunityMemberRolesBody:
      type: object
      required:
        - roleIds
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        roleIds:
          type: array
          description: Custom role ids to assign. The everyone role is implicit.
          items:
            type: string
...
    SignedPublicMutation:
      type: object
      description: Client-signed intent for a replicated public record. The node holds no user private keys, so the UI signs it. payloadDigest commits to the record without its proof; author.deviceCredential must be an authorized device of author.identityId in its current device authorization head. The winner of concurrent mutations is the higher sequence, then the lower digest; no wall clock is used.
      required:
        - version
        - operationId
        - kind
        - store
        - recordId
        - payloadDigest
        - predecessor
        - sequence
        - author
        - signature
      properties:
        version:
          type: integer
          enum: [1]
        operationId:
          type: string
        kind:
          type: string
          enum: [put, delete]
        store:
          type: string
        recordId:
          type: string
        payloadDigest:
          type: string
        predecessor:
          type: string
          nullable: true
          description: Digest of the previous mutation. Null if and only if sequence is 0.
        sequence:
          type: integer
          minimum: 0
        author:
          type: object
          required: [identityId, deviceCredential]
          properties:
            identityId:
              type: string
            deviceCredential:
              type: string
        signature:
          type: string
...